+++ wemake_python_styleguide/violations/refactoring.py
@@ -1053,6 +1053,17 @@ class ImplicitYieldFromViolation(ASTViolation):
         for item in some():
             yield item
 
+        # Correct:
+        yield from (
+            value[index:index + chunk_size]
+            for index in range(0, len(value), chunk_size)
+        )
+
+        # Wrong:
+        def chunks(value, chunk_size):
+            for index in range(0, len(value), chunk_size):
+                yield value[index:index + chunk_size]
+
     .. versionadded:: 0.13.0
 
     """